Windows: config refactoring
