+++ libcontainer/configs/config_test_unix.go
@@ -1,3 +1,5 @@
+// +build linux freebsd
+
 package configs
 
 import (
