+++ libcontainer/configs/config_test_windows.go
@@ -1,0 +1,3 @@
+package configs
+
+// All current tests are for Unix specific functionality
