+++ libcontainer/configs/config_windows.go
@@ -1,0 +1,8 @@
+package configs
+
+// Config defines configuration options for executing a process inside a contained environment.
+type Config struct {
+	BaseConfig
+
+	// Fields below here are platform specific
+}
